some protections in the pend vec && a overflow in BTC ex

+++ cpp09/ex00/BitcoinExchange.hpp
@@ -13,7 +13,7 @@
 class bitcoinExchange
 {
     private :
-        std::map<long int, double> Data;
+        std::map<unsigned long long int, double> Data;
         std::fstream rFile;
     public :
         bitcoinExchange(std::string file);
